+++ Editor/MergeAnimatorProcessor.cs
@@ -114,19 +114,27 @@ namespace nadena.dev.modular_avatar.core.editor
                 // We'll now continue processing the rest as normal.
             }
 
+            var writeDefaults = AnalyzeLayerWriteDefaults(controller);
+
+            foreach (var component in sorted)
+            {
+                MergeSingle(context, controller, component, writeDefaults);
+            }
+        }
+
+        internal static bool? AnalyzeLayerWriteDefaults(VirtualAnimatorController controller)
+        {
             bool? writeDefaults = null;
-            
-            var wdStateCounter = controller.Layers.SelectMany(l => l.StateMachine.AllStates())
+
+            var wdStateCounter = controller.Layers
+                .Where(l => !IsWriteDefaultsSafeLayer(l))
+                .SelectMany(l => l.StateMachine.AllStates())
                 .Select(s => s.WriteDefaultValues)
                 .GroupBy(b => b)
                 .ToDictionary(g => g.Key, g => g.Count());
 
             if (wdStateCounter.Count == 1) writeDefaults = wdStateCounter.First().Key;
-            
-            foreach (var component in sorted)
-            {
-                MergeSingle(context, controller, component, writeDefaults);
-            }
+            return writeDefaults;
         }
 
         private void MergeSingle(BuildContext context, VirtualAnimatorController targetController,
@@ -208,7 +216,7 @@ namespace nadena.dev.modular_avatar.core.editor
             Object.DestroyImmediate(merge);
         }
 
-        private bool IsWriteDefaultsSafeLayer(VirtualLayer virtualLayer)
+        private static bool IsWriteDefaultsSafeLayer(VirtualLayer virtualLayer)
         {
             if (virtualLayer.BlendingMode == AnimatorLayerBlendingMode.Additive) return true;
             var sm = virtualLayer.StateMachine;